pyproject.toml
[project]
name = "quantum-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "qiskit>=2.3.1",
    "qiskit-aer>=0.17.2",
    "qiskit-ibm-runtime>=0.46.1",
    "rustworkx>=0.17.1",
    "scipy>=1.17.1",
    "scipy-stubs~=1.17.1",
]
